Move cli.common to snowcli.api
