+++ src/snowcli/api/cli_global_context.py
