+++ src/snowcli/api/project_initialisation.py
@@ -17,8 +17,8 @@ def _create_project_template(template_name: str, project_directory: str):
 
 
 def add_init_command(app: Typer, project_type: str, template: str):
+    from snowcli.api.commands.decorators import global_options
     from snowcli.api.output.decorators import with_output
-    from snowcli.cli.common.decorators import global_options
 
     @app.command()
     @with_output
