+++ src/snowcli/api/utils/rendering.py
