+++ src/snowcli/app/main_typer.py
@@ -1,8 +1,8 @@
 import sys
 
 import click
-from snowcli.cli.common.cli_global_context import cli_context
-from snowcli.cli.common.flags import DEFAULT_CONTEXT_SETTINGS, DebugOption
+from snowcli.api.cli_global_context import cli_context
+from snowcli.api.commands.flags import DEFAULT_CONTEXT_SETTINGS, DebugOption
 from typer import Typer
