+++ src/snowcli/cli/containers/compute_pool/manager.py
@@ -1,4 +1,4 @@
-from snowcli.cli.common.sql_execution import SqlExecutionMixin
+from snowcli.api.sql_execution import SqlExecutionMixin
 from snowflake.connector.cursor import SnowflakeCursor
