+++ src/snowcli/cli/containers/services/commands.py
@@ -2,14 +2,14 @@ import sys
 from pathlib import Path
 
 import typer
+from snowcli.api.commands.decorators import global_options_with_connection
+from snowcli.api.commands.flags import DEFAULT_CONTEXT_SETTINGS
 from snowcli.api.output.decorators import with_output
 from snowcli.api.output.types import (
     CommandResult,
     QueryJsonValueResult,
     SingleQueryResult,
 )
-from snowcli.cli.common.decorators import global_options_with_connection
-from snowcli.cli.common.flags import DEFAULT_CONTEXT_SETTINGS
 from snowcli.cli.containers.common import print_log_lines
 from snowcli.cli.containers.services.manager import ServiceManager
