+++ src/snowcli/cli/exception_logging.py
@@ -1,7 +1,7 @@
 from logging import Logger
 from typing import Callable
 
-from snowcli.cli.common.cli_global_context import cli_context
+from snowcli.api.cli_global_context import cli_context
 
 
 def exception_logging(logger: Logger) -> Callable[[str, Exception], None]:
