+++ src/snowcli/cli/nativeapp/commands.py
@@ -2,15 +2,15 @@ import logging
 from typing import Optional
 
 import typer
-from snowcli.api.output.decorators import with_output
-from snowcli.api.output.types import CommandResult, MessageResult
-from snowcli.cli.common.cli_global_context import cli_context
-from snowcli.cli.common.decorators import (
+from snowcli.api.cli_global_context import cli_context
+from snowcli.api.commands.decorators import (
     global_options,
     global_options_with_connection,
     with_project_definition,
 )
-from snowcli.cli.common.flags import DEFAULT_CONTEXT_SETTINGS
+from snowcli.api.commands.flags import DEFAULT_CONTEXT_SETTINGS
+from snowcli.api.output.decorators import with_output
+from snowcli.api.output.types import CommandResult, MessageResult
 from snowcli.cli.nativeapp.common_flags import ForceOption, InteractiveOption
 from snowcli.cli.nativeapp.init import nativeapp_init
 from snowcli.cli.nativeapp.manager import NativeAppManager
