+++ src/snowcli/cli/nativeapp/manager.py
@@ -8,7 +8,7 @@ from textwrap import dedent
 from typing import Dict, List, Optional
 
 from snowcli.api.exception import SnowflakeSQLExecutionError
-from snowcli.cli.common.sql_execution import SqlExecutionMixin
+from snowcli.api.sql_execution import SqlExecutionMixin
 from snowcli.cli.connection.util import make_snowsight_url
 from snowcli.cli.nativeapp.artifacts import (
     ArtifactMapping,
