+++ src/snowcli/cli/nativeapp/version/commands.py
@@ -3,14 +3,14 @@ from typing import Optional
 
 import typer
 from click import MissingParameter
-from snowcli.api.output.decorators import with_output
-from snowcli.api.output.types import CommandResult, MessageResult, QueryResult
-from snowcli.cli.common.cli_global_context import cli_context
-from snowcli.cli.common.decorators import (
+from snowcli.api.cli_global_context import cli_context
+from snowcli.api.commands.decorators import (
     global_options_with_connection,
     with_project_definition,
 )
-from snowcli.cli.common.flags import DEFAULT_CONTEXT_SETTINGS
+from snowcli.api.commands.flags import DEFAULT_CONTEXT_SETTINGS
+from snowcli.api.output.decorators import with_output
+from snowcli.api.output.types import CommandResult, MessageResult, QueryResult
 from snowcli.cli.nativeapp.common_flags import ForceOption, InteractiveOption
 from snowcli.cli.nativeapp.policy import (
     AllowAlwaysPolicy,
