+++ src/snowcli/cli/object/__init__.py
@@ -1,5 +1,5 @@
 import typer
-from snowcli.cli.common.flags import DEFAULT_CONTEXT_SETTINGS
+from snowcli.api.commands.flags import DEFAULT_CONTEXT_SETTINGS
 from snowcli.cli.object.commands import app as show_app
 from snowcli.cli.object.stage.commands import app as stage_app
