+++ src/snowcli/cli/object/manager.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from click import ClickException
-from snowcli.cli.common.sql_execution import SqlExecutionMixin
+from snowcli.api.sql_execution import SqlExecutionMixin
 from snowcli.cli.constants import OBJECT_TO_NAMES, ObjectNames
 from snowflake.connector.cursor import SnowflakeCursor
