+++ src/snowcli/cli/object/stage/manager.py
@@ -6,8 +6,8 @@ from contextlib import nullcontext
 from pathlib import Path
 from typing import Optional, Union
 
+from snowcli.api.sql_execution import SqlExecutionMixin
 from snowcli.api.utils.path_utils import path_resolver
-from snowcli.cli.common.sql_execution import SqlExecutionMixin
 from snowcli.cli.project.util import to_string_literal
 from snowflake.connector.cursor import SnowflakeCursor
