+++ src/snowcli/cli/snowpark/commands.py
@@ -9,6 +9,16 @@ from typing import Dict, List, Set
 
 import typer
 from click import ClickException
+from snowcli.api.cli_global_context import cli_context
+from snowcli.api.commands.decorators import (
+    global_options,
+    global_options_with_connection,
+    with_project_definition,
+)
+from snowcli.api.commands.flags import (
+    DEFAULT_CONTEXT_SETTINGS,
+    execution_identifier_argument,
+)
 from snowcli.api.exception import (
     SecretsWithoutExternalAccessIntegrationError,
 )
@@ -19,19 +29,9 @@ from snowcli.api.output.types import (
     MessageResult,
     SingleQueryResult,
 )
+from snowcli.api.project_initialisation import add_init_command
 from snowcli.api.utils import file_utils
 from snowcli.api.utils.zipper import add_file_to_existing_zip
-from snowcli.cli.common.cli_global_context import cli_context
-from snowcli.cli.common.decorators import (
-    global_options,
-    global_options_with_connection,
-    with_project_definition,
-)
-from snowcli.cli.common.flags import (
-    DEFAULT_CONTEXT_SETTINGS,
-    execution_identifier_argument,
-)
-from snowcli.cli.common.project_initialisation import add_init_command
 from snowcli.cli.constants import DEPLOYMENT_STAGE, ObjectType
 from snowcli.cli.object.manager import ObjectManager
 from snowcli.cli.object.stage.manager import StageManager
