+++ src/snowcli/cli/snowpark/package/commands.py
@@ -4,10 +4,10 @@ import logging
 from pathlib import Path
 
 import typer
+from snowcli.api.commands.decorators import global_options_with_connection
+from snowcli.api.commands.flags import DEFAULT_CONTEXT_SETTINGS
 from snowcli.api.output.decorators import with_output
 from snowcli.api.output.types import CommandResult, MessageResult
-from snowcli.cli.common.decorators import global_options_with_connection
-from snowcli.cli.common.flags import DEFAULT_CONTEXT_SETTINGS
 from snowcli.cli.snowpark.package.manager import (
     cleanup_after_install,
     create,
