+++ src/snowcli/cli/snowpark/procedure_coverage/manager.py
@@ -8,7 +8,7 @@ import coverage
 import snowflake
 import typer
 from click import ClickException
-from snowcli.cli.common.sql_execution import SqlExecutionMixin
+from snowcli.api.sql_execution import SqlExecutionMixin
 from snowcli.cli.object.stage.manager import StageManager
 from snowflake.connector.cursor import SnowflakeCursor
