+++ src/snowcli/cli/sql/commands.py
@@ -2,9 +2,9 @@ from pathlib import Path
 from typing import Optional
 
 import typer
+from snowcli.api.commands.decorators import global_options_with_connection
 from snowcli.api.output.decorators import with_output
 from snowcli.api.output.types import CommandResult, MultipleResults, QueryResult
-from snowcli.cli.common.decorators import global_options_with_connection
 from snowcli.cli.sql.manager import SqlManager
 
 # simple Typer with defaults because it won't become a command group as it contains only one command
