+++ src/snowcli/cli/streamlit/commands.py
@@ -5,20 +5,20 @@ from typing import Optional
 import click
 import typer
 from click import ClickException
+from snowcli.api.cli_global_context import cli_context
+from snowcli.api.commands.decorators import (
+    global_options_with_connection,
+    with_experimental_behaviour,
+    with_project_definition,
+)
+from snowcli.api.commands.flags import DEFAULT_CONTEXT_SETTINGS
 from snowcli.api.output.decorators import with_output
 from snowcli.api.output.types import (
     CommandResult,
     MessageResult,
     SingleQueryResult,
 )
-from snowcli.cli.common.cli_global_context import cli_context
-from snowcli.cli.common.decorators import (
-    global_options_with_connection,
-    with_experimental_behaviour,
-    with_project_definition,
-)
-from snowcli.cli.common.flags import DEFAULT_CONTEXT_SETTINGS
-from snowcli.cli.common.project_initialisation import add_init_command
+from snowcli.api.project_initialisation import add_init_command
 from snowcli.cli.streamlit.manager import StreamlitManager
 
 app = typer.Typer(
