+++ tests_integration/conftest.py
@@ -12,8 +12,8 @@ from typing import Any, Dict, List, Optional
 
 import pytest
 import strictyaml
+from snowcli.api.cli_global_context import cli_context_manager
 from snowcli.app.cli_app import app
-from snowcli.cli.common.cli_global_context import cli_context_manager
 from snowcli.cli.project.definition import merge_left
 from strictyaml import as_document
 from typer import Typer
